toolkit/components/passwordmgr/test/mochitest/test_LoginManagerContent_passwordEditedOrGenerated.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test behavior of unmasking in LMC._passwordEditedOrGenerated</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="../../../satchel/test/satchel_common.js"></script>
  <script src="pwmgr_common.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<script>
let readyPromise = registerRunTests(1);

let origin = window.location.origin;
addLoginsInParent(
  [origin, "https://autofill", null, "user1", "pass1"],
  [origin, "https://autofill", null, "user2", "pass2"]);

function preventDefaultAndStopProgagation(event) {
  event.preventDefault();
  event.stopImmediatePropagation();
}
</script>
  <p id="display">
    <form id="form1" action="https://autofill" onsubmit="return false;">
      <input type="text" name="uname">
      <input type="password" name="pword">
      <button type="submit">Submit</button>
    </form>
  </p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
<script>
const { LoginManagerChild } = SpecialPowers.ChromeUtils.import(
  "resource://gre/modules/LoginManagerChild.jsm"
);

add_setup(async () => {
  await SpecialPowers.pushPrefEnv({"set": [
    ["signon.generation.available", true],
    ["signon.generation.enabled", true],
  ]});
  await readyPromise;
});

add_task(async function test_preventDefaultAndStopPropagation() {
  let pword = getFormElementByName(1, "pword");
  pword.focus();

  pword.addEventListener("focus", preventDefaultAndStopProgagation);
  pword.addEventListener("focus", preventDefaultAndStopProgagation, true);
  pword.addEventListener("blur", preventDefaultAndStopProgagation);
  pword.addEventListener("blur", preventDefaultAndStopProgagation, true);

  SpecialPowers.wrap(pword).setUserInput("generatedpass");
  LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "Before first fill");
  LoginManagerChild.forWindow(window)._passwordEditedOrGenerated(pword, { triggeredByFillingGenerated: true });
  LOGIN_FIELD_UTILS.checkPasswordMasked(pword, false, "After first fill");
  synthesizeKey("KEY_Tab"); // blur
  LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "After blur");
  synthesizeKey("KEY_Tab", { shiftKey: true }); // focus again
  LOGIN_FIELD_UTILS.checkPasswordMasked(pword, false, "After shift-tab to focus again");

  // Cleanup
  recreateTree(document.getElementById("form1"));
  await promiseFormsProcessedInSameProcess();
});

add_task(async function test_fieldsMaskedAfterSavedLoginFill() {
  let pword = getFormElementByName(1, "pword");
  pword.focus();

  SpecialPowers.wrap(pword).setUserInput("generatedpass");
  LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "Before first fill");
  LoginManagerChild.forWindow(window)._passwordEditedOrGenerated(pword, { triggeredByFillingGenerated: true });
  LOGIN_FIELD_UTILS.checkPasswordMasked(pword, false, "After first fill");
  synthesizeKey("KEY_Tab", { shiftKey: true }); // blur pw, focus un
  LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "After blur");

  info("Filling username matching saved login");
  sendString("user1");

  let processedPromise = promiseFormsProcessedInSameProcess();
  synthesizeKey("KEY_Tab"); // focus again and trigger a fill of the matching password
  await processedPromise;
  is(pword.value, "pass1", "Saved password was filled")
  LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "After filling a saved login");

  // Cleanup
  recreateTree(document.getElementById("form1"));
  await promiseFormsProcessedInSameProcess();
});

add_task(async function test_fieldsMaskedAfterReplacingWholeValue() {
  let pword = getFormElementByName(1, "pword");
  pword.focus();

  SpecialPowers.wrap(pword).setUserInput("generatedpass");
  LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "Before first fill");
  LoginManagerChild.forWindow(window)._passwordEditedOrGenerated(pword, { triggeredByFillingGenerated: true });
  LOGIN_FIELD_UTILS.checkPasswordMasked(pword, false, "After first fill");
  synthesizeKey("KEY_Tab", { shiftKey: true }); // blur pw, focus un
  LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "After blur");

  synthesizeKey("KEY_Tab"); // focus again and replace the whole password value
  info("Replacing password field value with arbitrary string");
  sendString("some_other_password");
  is(pword.value, "some_other_password", "Whole password replaced")
  LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "Replaced password value");

  synthesizeKey("KEY_Tab"); // blur pw
  LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "After blur");
  synthesizeKey("KEY_Tab", { shiftKey: true }); // focus pw again
  LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "After focus again");

  // Cleanup
  recreateTree(document.getElementById("form1"));
  await promiseFormsProcessedInSameProcess();
});

add_task(async function test_fieldsUnmaskedAfterAddingCharacter() {
  let pword = getFormElementByName(1, "pword");
  pword.focus();

  SpecialPowers.wrap(pword).setUserInput("generatedpass");
  LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "Before first fill");
  LoginManagerChild.forWindow(window)._passwordEditedOrGenerated(pword, { triggeredByFillingGenerated: true });
  LOGIN_FIELD_UTILS.checkPasswordMasked(pword, false, "After first fill");
  synthesizeKey("KEY_Tab", { shiftKey: true }); // blur pw, focus un
  LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "After blur");

  synthesizeKey("KEY_Tab"); // focus again
  synthesizeKey("KEY_ArrowRight"); // Remove the selection
  info("Adding a character to the end of the password");
  sendString("@");
  is(pword.value, "generatedpass@", "Character was added to the value")
  LOGIN_FIELD_UTILS.checkPasswordMasked(pword, false, "Added @");

  synthesizeKey("KEY_Tab"); // blur pw
  LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "After blur after @");
  synthesizeKey("KEY_Tab", { shiftKey: true }); // focus pw again
  LOGIN_FIELD_UTILS.checkPasswordMasked(pword, false, "After focus after @");

  // Cleanup
  recreateTree(document.getElementById("form1"));
  await promiseFormsProcessedInSameProcess();
});

add_task(async function test_typeNotPassword() {
  let pword = getFormElementByName(1, "pword");
  pword.focus();

  LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "Before first fill");
  SpecialPowers.wrap(pword).setUserInput("generatedpass");
  LoginManagerChild.forWindow(window)._passwordEditedOrGenerated(pword, { triggeredByFillingGenerated: true });
  LOGIN_FIELD_UTILS.checkPasswordMasked(pword, false, "After first fill");

  // Simulate a website doing their own unmasking and re-masking
  pword.type = "text";
  await new Promise(resolve => SimpleTest.executeSoon(resolve));
  pword.type = "password";

  synthesizeKey("KEY_Tab"); // blur
  LOGIN_FIELD_UTILS.checkPasswordMasked(pword, true, "After blur");
  synthesizeKey("KEY_Tab", { shiftKey: true }); // focus again
  LOGIN_FIELD_UTILS.checkPasswordMasked(pword, false, "After shift-tab to focus again");

  // Cleanup
  recreateTree(document.getElementById("form1"));
  await promiseFormsProcessedInSameProcess();
});
</script>
</body>
</html>
